Added timing testing for all code in test.py, run_all.py. test.py allows the user to run specific sp, while run_all tests all.
